Move CatkinResourceRetriever to io namespace
